# [Testing] Init basic performance tests, starting with dashboard