--- a/project/npda/views/dashboard/partials.py
+++ b/project/npda/views/dashboard/partials.py
@@ -575,7 +575,7 @@ def get_simple_bar_chart_pcts_partial(request):
             {"chart_html": chart_html},
         )
     except Exception as e:
-        logger.error("Error generating simple bar chart pcts", exc_info=True)
+        logger.error(f"Error generating simple bar chart pcts {e=}", exc_info=True)
         return render(
             request,
             "dashboard/simple_bar_chart_pcts_partial.html",