--- a/pytest.ini
+++ b/pytest.ini
@@ -9,7 +9,9 @@ python_files = test_*.py
 # RE USE TEST DB AS DEFAULT
 addopts = 
     --reuse-db
+    # Don't run these by default
     -k "not examples"
+    -m "not performance"
 
 # ENABLE LOGGING TO CONSOLE
 log_cli = true
@@ -18,5 +20,6 @@ log_cli_level = INFO
 markers =
     examples: mark test as workshop-type / example test
     seed: mark test as 'meta test', just used for seeding
+    performance: mark test as performance test
 
 asyncio_mode = auto