Signed-off-by: anchit-chandran <anchit97123@gmail.com>

### Overview

Inits some basic performance tests, starting with the dashboard view.

NOTE: these are IGNORED by default (using `@pytest.mark.performance`). Use the bash cmd to run performance tests, as mentioned updated Testing docs section:

```
s/test -m performance -s  
```

The added test is completely machine-dependent, but worth as a starting point.

`test_performance.py::test_dashboard_view_response_time`'s overall structure is:

1. Seed `N_PATIENTS` with `VISIT_TYPES` in the logged in user's PDU
2. Start the request for the `dashboard` view (this is the top level view whose template makes multiple hx-gets to each viz's card.
4. Use simple `time.time()` to measure duration
3. Use `beautifulSoup` to parse the `dashboard` view's rendered template to capture each of these `hx-get` urls and associated data in `hx-vals` dynamically
4. For each of these, ping off requests to each view

A final report is logged:

```
s/test -m performance -s                                                                                        
...

---------------------------------------------------------------------------------------------------------- live log call -----------------------------------------------------------------------------------------------------------
INFO     test_performance:test_performance.py:63 Seeding 1000 patients with 11 visit types...
INFO     test_performance:test_performance.py:74 DONE!
INFO     test_performance:test_performance.py:75 Creating submission...
INFO     test_performance:test_performance.py:86 DONE!
INFO     test_performance:test_performance.py:90        ⏳ Getting top-level dashboard view response...
INFO     test_performance:test_performance.py:94 ***********************N_PATIENTS: 1000 * VISIT_TYPES: 11***********************
INFO     test_performance:test_performance.py:101       Dashboard view response time: 4.61618709564209 seconds.
INFO     test_performance:test_performance.py:178       ⏳ Rendering htmx partials...
         ✅ /dashboard response in 4.725 seconds
         ✅ /get_progress_bar_chart_partial response in 0.022 seconds
         ✅ /get_progress_bar_chart_partial response in 0.006 seconds
         ✅ /get_progress_bar_chart_partial response in 0.007 seconds
         ✅ /get_simple_bar_chart_pcts_partial?color=E00087 response in 0.088 seconds
         ✅ /get_simple_bar_chart_pcts_partial?color=E00087 response in 0.012 seconds
         ✅ /get_waffle_chart_partial response in 0.022 seconds
         ✅ /get_hcl_scatter_plot response in 0.014 seconds
         ✅ /get_simple_bar_chart_pcts_partial response in 0.012 seconds
         ✅ /get_simple_bar_chart_pcts_partial?color=E00087 response in 0.013 seconds
         ✅ /get_waffle_chart_partial response in 0.019 seconds
         ✅ /get_treemap_chart_partial response in 0.010 seconds
         ✅ /get_waffle_chart_partial response in 0.018 seconds
         ✅ /get_map_chart_partial response in 7.214 seconds
         ✅ /get_patient_level_report_partial response in 2.844 seconds
         ✅ /get_patient_level_report_partial response in 0.929 seconds
         ✅ /get_patient_level_report_partial response in 0.031 seconds
         ✅ /get_patient_level_report_partial response in 2.783 seconds
         ✅ /get_patient_level_report_partial response in 8.773 seconds
INFO     test_performance:test_performance.py:193 
        SORTED dashboard view request times:
INFO     test_performance:test_performance.py:195       ⏳ /get_patient_level_report_partial       : 8.773 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_map_chart_partial                  : 7.214 seconds
INFO     test_performance:test_performance.py:195       ⏳ /dashboard                              : 4.725 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_patient_level_report_partial       : 2.844 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_patient_level_report_partial       : 2.783 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_patient_level_report_partial       : 0.929 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_simple_bar_chart_pcts_partial      : 0.088 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_patient_level_report_partial       : 0.031 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_waffle_chart_partial               : 0.022 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_progress_bar_chart_partial         : 0.022 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_waffle_chart_partial               : 0.019 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_waffle_chart_partial               : 0.018 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_hcl_scatter_plot                   : 0.014 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_simple_bar_chart_pcts_partial      : 0.013 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_simple_bar_chart_pcts_partial      : 0.012 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_simple_bar_chart_pcts_partial      : 0.012 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_treemap_chart_partial              : 0.010 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_progress_bar_chart_partial         : 0.007 seconds
INFO     test_performance:test_performance.py:195       ⏳ /get_progress_bar_chart_partial         : 0.006 seconds
INFO     test_performance:test_performance.py:196       Total Page Load Time: 32.159 seconds
PASSED
```

### Code changes

- Adds tests
- Updates `pytest.ini` to describe `performance` marker

### Documentation changes (done or required as a result of this PR)
- Updates Testing section to mention how to run performance tests

### Related Issues
#538